samples/python/foundry-local/integrate-sdks/requirements.txt
openai
foundry-local-sdk
requests
